fix(stella-cli): split the run ledger out of self_driving_cmd.rs so main compiles its own ceiling
`crates/stella-cli/src/self_driving_cmd.rs` reached 1501 lines on the merged
tree and is not grandfathered, so `file-size` failed on `main` itself and —
per AGENTS.md's `main-red-hold` — reddened every open PR with a failure none
of their authors caused.

The 1500-line ceiling is not the defect; the file is. `run_start`, `run_end`,
`run_end_as` and `runs_report` are the run ledger: one bracket around the
cycles `lifecycle` drives, sharing no state with the argument tree or the
cycle reporting around them, and called from exactly four sites in one match
arm. They move to `self_driving_cmd/run_lifecycle.rs` beside the fourteen
submodules already there, following `stella-core`'s `driver/settlement.rs`.

The move is mechanical: no body changed, only the visibility (`pub(super)`)
and the four call sites. `BTreeMap` left the parent's import list with them.

No entry was added to `scripts/file-size-baseline.txt` — the baseline covers
files that predate the guard, and a crossing bought with a raised ceiling is
the expedient CLAUDE.md forbids. The parent lands at 1421 lines, 79 under.

Closes #4044
